fix comm
