+++ src/frontend/src/error.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use core::convert::Infallible;
-use core::ops::FromResidual;
-
 use risingwave_batch::error::BatchError;
 use risingwave_common::array::ArrayError;
 use risingwave_common::error::{BoxedError, NoFunction, NotImplemented};
@@ -255,37 +252,5 @@ impl From<BoxedError> for RwError {
         // This is essentially expanded from `anyhow::anyhow!(e)`.
         let e = anyhow::__private::kind::BoxedKind::anyhow_kind(&e).new(e);
         ErrorCode::Uncategorized(e).into()
-    }
-}
-
-/// Result when applying an optimization rule.
-pub enum OResult<T> {
-    /// Successfully optimized the input.
-    Ok(T),
-    /// The current rule is not applicable to the input.
-    /// The caller may try other rules.
-    NotApplicable,
-    /// There was an unrecoverable error while applying the rule.
-    /// The caller should stop trying other rules and report the error.
-    Err(RwError),
-}
-
-/// `Option<T>` -> `OResult<T>`, `None` -> `NotApplicable`
-impl<T> FromResidual<Option<Infallible>> for OResult<T> {
-    fn from_residual(residual: Option<Infallible>) -> Self {
-        match residual {
-            Some(_) => unreachable!(),
-            None => Self::NotApplicable,
-        }
-    }
-}
-
-/// `Result<T>` -> `OResult<T>`
-impl<T> FromResidual<Result<Infallible>> for OResult<T> {
-    fn from_residual(residual: Result<Infallible>) -> Self {
-        match residual {
-            Ok(_) => unreachable!(),
-            Err(e) => Self::Err(e),
-        }
     }
 }
