+++ src/frontend/src/optimizer/rule/mod.rs
@@ -14,8 +14,7 @@
 
 //! Define all [`Rule`]
 
-use super::PlanRef;
-use crate::error::OResult;
+use super::{OResult, PlanRef};
 
 /// A one-to-one transform for the [`PlanNode`](super::plan_node::PlanNode), every [`Rule`] should
 /// downcast and check if the node matches the rule.
